move to /src/pkg
